backend/requirements.txt
flask
flask-cors
python-chess
pygame
gunicorn
